# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Desafio aiqfome",
        "description": "Documentação da API de clientes e favoritos",
        "contact": {
            "email": "exemple@exemple.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/clients/{client}/favorites": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get client's favorite products",
                "description": "Returns a list of products favorited by the specified client.",
                "operationId": "a47f6adf09fa03b5e3d8738060e89aa3",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of client's favorite products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Favorites retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/{client}/favorites/{product_api_id}": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add product to client's favorites",
                "operationId": "ae25812ef879224c0fdf81dd7e85ef7b",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_api_id",
                        "in": "path",
                        "description": "External product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product added to favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found in external API"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Remove product from client's favorites",
                "operationId": "446851d08b22cea7a803d8112235e5f3",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_api_id",
                        "in": "path",
                        "description": "External product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product removed from favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Product is not in client's favorites"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [
                    "product_api_id",
                    "title",
                    "image",
                    "price",
                    "review"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_api_id": {
                        "type": "string",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "Smartphone XYZ"
                    },
                    "image": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/images/product.jpg"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1299.99
                    },
                    "review": {
                        "type": "string",
                        "example": "Excellent performance and battery life."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-25T14:56:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-25T15:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}